Cast size to int

+++ include/LightGBM/dataset.h
@@ -714,10 +714,10 @@ class Dataset {
     if (static_cast<int>(raw_data_.size()) > num_numeric_features_) {
       raw_data_.resize(num_numeric_features_);
     }
-    for (size_t i = 0; i < raw_data_.size(); ++i) {
+    for (size_t i = 0; i < static_cast<int>(raw_data_.size()); ++i) {
       raw_data_[i].resize(num_rows);
     }
-    int curr_size = raw_data_.size();
+    int curr_size = static_cast<int>(raw_data_.size());
     for (int i = curr_size; i < num_features_; ++i) {
       int feat_ind = numeric_feature_map_[i];
       if (feat_ind >= 0) {
